requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Pydantic for data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# LangGraph for workflow orchestration
# Using compatible versions to avoid ImportError
langgraph==0.2.34
langchain-core==0.3.10
langchain==0.3.3

# Groq API client
groq==0.11.0

# Environment variable management
python-dotenv==1.0.1

# HTTP client
httpx==0.27.2

# JSON processing
orjson==3.10.7

# CORS and middleware
starlette==0.38.6

# Type hints support
typing-extensions==4.12.2

